The path https://yandex.ru/games/sdk/v2 is outdated.
https://yandex.com/dev/games/doc/en/sdk/sdk-about

Alert

The path https://yandex.ru/games/sdk/v2 is outdated. To ensure your game passes moderation successfully, specify the up-to-date path for connecting Yandex Games SDK depending on the placement method:

If you upload the game archive in the developer console, specify a relative path. This is the recommended option.

If you are using integration via an iframe, specify an absolute path.